requirements.txt
absl-py==0.9.0
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.5
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.1
boto3==1.12.39
botocore==1.15.39
cachetools==4.0.0
cassandra-driver==3.22.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
confluent-kafka==1.3.0
cryptography==2.8
cycler==0.10.0
data==0.4
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
docutils==0.15.2
dse==3.3.0
dse-driver==2.11.1
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
Flask-MySQLdb==0.2.0
funcsigs==1.0.2
future==0.18.2
geomet==0.1.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
grpcio==1.27.2
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipython-unittest==0.3.1
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kafka-python==2.0.1
kiwisolver==1.1.0
latex==0.7.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydicom==1.4.2
Pygments==2.6.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.6
pypng==0.0.20
pyrsistent==0.15.7
python-dateutil==2.8.1
pytorch-lightning==0.7.1
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
shutilwhich==1.1.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
sshtunnel==0.1.5
tempdir==0.7.1
tensorboard==2.1.1
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.10
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==3.1.0